simple_voice_agent/requirements.txt
livekit>=0.9.0
livekit-agents>=0.7.0
livekit-plugins-google
fastapi
uvicorn
python-dotenv
